sending file

+++ src/screens/Chat/Content.js
@@ -34,12 +34,14 @@ import {
   removePreloader,
   replyMessage,
   setCurrentRoomId,
+  setSendingMessages,
 } from '../../actions/messageActions'
 import { d_message } from '../../constants/api'
 import _ from 'lodash'
 import * as ICONS from '../../assets/icons'
 import Loader from '../../common/Loader'
 import RNPermissions from 'react-native-permissions'
+import { socket } from '../../utils/socket'
 // import RNPermissions from 'react-native-permissions'
 
 const {
@@ -94,9 +96,12 @@ class Content extends Component {
       uploadMessages,
       dialog,
       editedMessage2,
+      sendingMessages,
     } = this.props
-
     let { messages } = this.props
+
+    const companyKey = dialog.company
+    const dialogKey = dialog._id
 
     if (!messages.length && Object.keys(dialog).length) {
       this.props.setMessages(dialog.messages)
@@ -117,7 +122,6 @@ class Content extends Component {
           messages[messageIndex].text = editedMessage2.text
           messages[messageIndex].edited = editedMessage2.edited
         }
-
         this.props.setMessages(messages)
         this.props.setEditedMessage(null)
       }
@@ -128,7 +132,15 @@ class Content extends Component {
       ? scrolledMessages
       : messages
 
-    let reversedMessages = [...currentMessages].sort(
+    const newSendingMessages = !!(
+      sendingMessages[companyKey] &&
+      sendingMessages[companyKey][dialogKey] &&
+      sendingMessages[companyKey][dialogKey].messages
+    )
+      ? sendingMessages[companyKey][dialogKey].messages
+      : []
+
+    let reversedMessages = [...currentMessages, ...newSendingMessages].sort(
       (x, y) => new Date(y.created_at) - new Date(x.created_at),
     )
 
@@ -645,25 +657,47 @@ class Content extends Component {
   }
 
   deleteMessage = currentMessage => {
-    const { currentRoomId, dialogs, dialog } = this.props
+    const {
+      currentRoomId,
+      dialogs,
+      dialog,
+      setSendingMessages,
+      sendingMessages,
+    } = this.props
     let { messages } = this.props
-    sendRequest({
-      r_path: d_message,
-      method: 'delete',
-      attr: {
-        dialog_id: currentRoomId,
-        messages: [currentMessage._id],
-      },
-      success: res => {},
-      failFunc: err => {},
-    })
-    messages = messages.filter(message => message._id !== currentMessage._id)
-    const index = dialogs.findIndex(d => d._id === dialog._id)
-    dialogs[index].messages = dialogs[index].messages.filter(
-      m => m._id !== currentMessage._id,
-    )
-    setDialogs(dialogs)
-    this.props.setMessages(messages)
+    if (currentMessage.myMessage && currentMessage.failed) {
+      const companyKey = dialog.company
+      const dialogKey = dialog._id
+      let receivedMessages = { ...sendingMessages }
+      receivedMessages[companyKey][dialogKey].messages = receivedMessages[
+        companyKey
+      ][dialogKey].messages.filter(m => m._id !== currentMessage._id)
+      setSendingMessages(receivedMessages)
+    } else {
+      sendRequest({
+        r_path: d_message,
+        method: 'delete',
+        attr: {
+          dialog_id: currentRoomId,
+          messages: [currentMessage._id],
+        },
+        success: res => {},
+        failFunc: err => {},
+      })
+      messages = messages.filter(message => message._id !== currentMessage._id)
+      const index = dialogs.findIndex(d => d._id === dialog._id)
+      dialogs[index].messages = dialogs[index].messages.filter(
+        m => m._id !== currentMessage._id,
+      )
+      setDialogs(dialogs)
+      this.props.setMessages(messages)
+    }
+  }
+
+  repeatMessage = message => {
+    // todo date dnel ev hin@ jnjel amenanerqevic avelacnel
+    const { currentRoom } = this.props
+    socket.emit('message', { receiver: currentRoom, message: message.text })
   }
 
   _scrollToBottom = () => {
@@ -777,49 +811,59 @@ class Content extends Component {
     const { user, dialog } = this.props
     this.props.setCurrentRoomId(dialog._id)
     let actions = []
-    if (
-      message._id &&
-      message.type === 'text' &&
-      message.sender._id === user._id
-    ) {
+    if (!message.myMessage) {
+      if (
+        message._id &&
+        message.type === 'text' &&
+        message.sender._id === user._id
+      ) {
+        actions.push({
+          title: 'Сделать задачей',
+          action: () => this.turnToTask(message),
+        })
+        actions.push({
+          title: 'Редактировать',
+          action: () => this.editMessage(message),
+        })
+      }
+      if (message._id && message.type === 'text') {
+        actions.push({
+          title: 'Копировать',
+          action: () => this.copyMessage(message),
+        })
+      }
       actions.push({
-        title: 'Сделать задачей',
-        action: () => this.turnToTask(message),
+        title: 'Ответить',
+        action: () => this.replyMessage(message),
       })
       actions.push({
-        title: 'Редактировать',
-        action: () => this.editMessage(message),
+        title: 'Переслать',
+        action: () => this.forwardMessage(message),
       })
-    }
-    if (message._id && message.type === 'text') {
-      actions.push({
-        title: 'Копировать',
-        action: () => this.copyMessage(message),
-      })
-    }
-    actions.push({
-      title: 'Ответить',
-      action: () => this.replyMessage(message),
-    })
-    actions.push({
-      title: 'Переслать',
-      action: () => this.forwardMessage(message),
-    })
 
-    if (
-      message._id &&
-      (message.type === 'file' ||
-        message.type === 'video' ||
-        message.type === 'image')
-    ) {
+      if (
+        message._id &&
+        (message.type === 'file' ||
+          message.type === 'video' ||
+          message.type === 'image')
+      ) {
+        actions.push({
+          title: 'Сохранить',
+          action: () => {
+            this.download(message)
+          },
+        })
+      }
+    } else if (message.myMessage && message.failed) {
       actions.push({
-        title: 'Сохранить',
-        action: () => {
-          this.download(message)
-        },
+        title: 'Повторить',
+        action: () => this.repeatMessage(message),
       })
     }
-    if (message._id && message.sender._id === user._id) {
+    if (
+      (message.myMessage && message.failed) ||
+      !!(message._id && message.sender._id === user._id)
+    ) {
       actions.push({
         title: 'Удалить',
         action: () => this.messageDeleteConfirmation(message),
@@ -881,6 +925,7 @@ const mapStateToProps = state => ({
   dialog: state.dialogsReducer.dialog,
   currentDialog: state.dialogsReducer.currentDialog,
   message: state.messageReducer.message,
+  sendingMessages: state.messageReducer.sendingMessages,
 })
 
 const mapDispatchToProps = dispatch => ({
@@ -893,5 +938,6 @@ const mapDispatchToProps = dispatch => ({
   replyMessage: _ => dispatch(replyMessage(_)),
   setCurrentRoomId: _ => dispatch(setCurrentRoomId(_)),
   setEditedMessage: _ => dispatch(getEditedMessage(_)),
+  setSendingMessages: _ => dispatch(setSendingMessages(_)),
 })
 export default connect(mapStateToProps, mapDispatchToProps)(Content)
